refactor(bft): move bft tests into separate file
